refactor(netlist): extract the builder -- netlist.py is now a pure facade
The graph->model builder (build_netlist, build_netlist_from_graph,
index_module, and all their helpers) moves to a new netlist_build.py (3,309
lines); component_line joins the deprecated renderer. netlist.py collapses to a
176-line facade that re-exports the public API from all six sibling modules, so
every importer (cli, diff, mcp, describe, tests) is unchanged. diff.py's one
private-helper import (_selector_lv_type) repointed to netlist_build.

This completes the netlist decomposition: the former 6,325-line god-module is
now seven single-responsibility modules -- facade / models / grammar /
render_lvnet / json / deprecated-renderer / build. The builder's own internal
sub-decomposition is deferred to the dedicated god-module refactor.

Byte-identical lvnet; pyright src/ clean; 207 tests green; one-way deps, no
import cycle.
